Update Table component
