+++ src/Table/index.ts
@@ -1,5 +1,11 @@
+import ReactTable from 'react-table';
+import treeTableHOC from 'react-table/lib/hoc/treeTable';
+import selectTableHOC from 'react-table/lib/hoc/selectTable';
+
 import { Table } from './Table.component';
 
-export { Table, Props as TableProps } from './Table.component';
+const TreeTable = treeTableHOC(ReactTable);
+const SelectTable = selectTableHOC(ReactTable);
+const SelectTreeTable = selectTableHOC(treeTableHOC(ReactTable));
 
-export default Table;
+export { TreeTable, SelectTable, SelectTreeTable, Table };
